Change 'in' (and '!in') filter to work with a hash for performance.
The 'in' filter was previously chaining if statements leading to
O(n*m) running time if using the returned filter function while
iterating over a list of n features with m items in the filter.

This change puts the 'in' arguments into a simple javascript hash
for O(n) time.

To prevent perf regressions for small filters we use the original code
path below a hard-coded threshold (currently set at 30 based on
rudimentary perf testing).